// packages/chai-contents-view/src/types/templateContents.ts
export type IconTextContent = {
  type: "iconText";
  data: {
    icon: {
      src: string;
    };
    text: string;
  }[];
};
// ...
export type AudioRecordContent = {
  type: "audioRecord";
  data: {
    audio: {
      src: string;
    };
  }[];
};

export type DragAndDropContent = {
  type: "dragAndDrop";
  data: {
    choices: string[];
    answerIndex: number;
    explanation: Explanation;
  }[];
};

export type StudyWordsContent = {
  type: "studyWords";
  data: {
    text: string; // html
    pronunciation: string; // html
    meaning: string; // html
    audio: {
      src: string;
    };
  }[];
};

export type DialogContent = {
  type: "dialog";
  data: DialogData[];
};

export type DialogData = {
  id: ID; // 대화를 구분하기 위한 id - 같은 사람이 여러번 말할 경우 구분하기 위함
  icon: {
    src: string;
  };
  text: string;
  pronunciation?: string;
  meaning?: string;
  hasQuestion: boolean;
  question?: {
    choices?: string[];
    answerIndex?: number;
    multiChoices?: MultiChoice[];
  };
  audio?: {
    src: string;
  };
};

export type MultiChoice = {
  text: string;
  pronunciation: string;
  answerIndex: number;
};